Testando csrf_exempt

+++ tasks/views.py
@@ -5,7 +5,7 @@ from django.core import serializers
 from tasks.serializer import TaskSerializer
 from rest_framework.parsers import JSONParser
 from django.forms.models import model_to_dict
-from rest_framework.decorators import api_view
+from django.views.decorators.csrf import csrf_exempt
 
 
 
@@ -13,22 +13,25 @@ from rest_framework.decorators import api_view
 # def index(request):
 #     return HttpResponse("Hello, world. You're at the tasks index.")
 
-@api_view(["GET"])
+@csrf_exempt
 def get_all_tasks(request):
+    if request.method == 'GET':
         tasks = Task.objects.all()
         tasks_list = serializers.serialize("json",tasks)
         return HttpResponse(tasks_list,content_type="application/json")
 
-@api_view(["GET"]) 
+@csrf_exempt
 def get_task(request,task_id):
+    if request.method == 'GET':
         try:
             task = Task.objects.get(pk=task_id)
             return JsonResponse(model_to_dict(task),safe=False)
         except:
             return Http404("Id não encontrado")
         
-@api_view(["PUT"])
+@csrf_exempt
 def edit_task(request,task_id):
+    if request.method == 'PUT':
         try:
             task = Task.objects.get(pk=task_id)
             request_data = JSONParser().parse(request)
@@ -41,8 +44,9 @@ def edit_task(request,task_id):
             return Http404("Id não encontrado")
             
             
-@api_view(["DELETE"])
+@csrf_exempt
 def delete_task(request, task_id):
+    if request.method == "DELETE":
         try:
             task = Task.objects.get(pk=task_id)
             task.delete()
@@ -51,8 +55,9 @@ def delete_task(request, task_id):
         except:
             return Http404("Id não encontrado")
         
-@api_view(["POST"])
+@csrf_exempt
 def create_task(request):
+    if request.method == "POST":
         request_data = JSONParser().parse(request)
         task = TaskSerializer(data=request_data)
         if task.is_valid():
